WIP: make `pallet-roles` compilable
